+++ .settings/org.eclipse.core.resources.prefs
@@ -1,3 +1,4 @@
 eclipse.preferences.version=1
-encoding//src/kr/co/bttf/app/board/FrontController.java=UTF-8
+encoding//src/kr/co/bttf/app/board/FrontControllerCss.java=UTF-8
+encoding//src/kr/co/bttf/app/member/MemberFrontController.java=UTF-8
 encoding/<project>=UTF-8
